+++ doc/changelog.rst
@@ -123,6 +123,9 @@ Breaking Changes in 4.0
   :class:`~bson.dbref.DBRef`.
 - The "tls" install extra is no longer necessary or supported and will be
   ignored by pip.
+- ``tz_aware``, an argument for :class:`~bson.json_util.JSONOptions`,
+  now defaults to ``False`` instead of ``True``. ``json_util.loads`` now
+decodes datetime as naive by default.
 - ``directConnection`` URI option and keyword argument to :class:`~pymongo.mongo_client.MongoClient`
   defaults to ``False`` instead of ``None``, allowing for the automatic
   discovery of replica sets. This means that if you
@@ -132,6 +135,7 @@ Breaking Changes in 4.0
   with :meth:`~pymongo.collection.Collection.find`.
 - ``name`` is now a required argument for the :class:`pymongo.driver_info.DriverInfo` class.
 
+ d
 Notable improvements
 ....................